add new func updateExistingConfigFromFlags
